Update to v0.2.1
